Major reorganization: separating out components relating to a single card to /components/card (#204)
* Major reorganization: separating out components relating to a single card to /components/card

* minor
