+++ .codeclimate.yml
@@ -50,6 +50,6 @@ exclude_paths:
   - test/
   - src/common/components/img/
   # The following are by-and-large not our own code, so it's not worth analyzing them:
-  - src/common/components/game/Sprite.js
+  - src/common/components/Sprite.js
   - src/common/components/react-hexgrid/
   - src/common/components/react-textfit/
