+++ src/common/components/Sprite.js
@@ -3,7 +3,7 @@
 import React, { PureComponent } from 'react';
 import { number, string } from 'prop-types';
 
-import { hashCode, inBrowser } from '../../util/common';
+import { hashCode, inBrowser } from '../util/common';
 
 export default class Sprite extends PureComponent {
   static propTypes = {
