+++ src/common/components/card/Identicon.js
