+++ src/common/components/card/Sentence.js
