+++ src/common/components/card/Word.js
